pyproject.toml
[project]
name = "ovh-dynhost-updater"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "rich>=14.2.0",
]
